4.6 Divide and Conquer
