+++ users/createAccount/createAccount.resolvers.js
@@ -1,0 +1,50 @@
+import bcrypt from "bcrypt";
+import client from "../../client.mjs";
+
+export default {
+  Mutation: {
+    createAccount: async (
+      _,
+      { firstName, lastName, username, email, password }
+    ) => {
+      try {
+        // check if username or email are already on DB.
+        const existingUser = await client.user.findFirst({
+          where: {
+            OR: [
+              {
+                username,
+              },
+              {
+                email,
+              },
+            ],
+          },
+        });
+        if (existingUser) {
+          throw new Error("This username/password is already taken.");
+        }
+
+        // hash password
+        const uglyPassword = await bcrypt.hash(password, 10);
+        console.log(uglyPassword);
+
+        // save the user
+        const user = await client.user.create({
+          data: {
+            username,
+            email,
+            firstName,
+            lastName,
+            password: uglyPassword,
+          },
+        });
+        return {
+          ok: true,
+        };
+      } catch (e) {
+        return e;
+      }
+    },
+  },
+};
