+++ users/login/login.resolvers.js
@@ -1,11 +1,10 @@
 import bcrypt from "bcrypt";
 import jwt from "jsonwebtoken";
-import client from "../client.mjs";
+import client from "../../client.mjs";
 
 export default {
   Mutation: {
     login: async (_, { username, password }) => {
-      // find user with args.username
       const user = await client.user.findFirst({ where: { username } });
       if (!user) {
         return {
@@ -13,16 +12,13 @@ export default {
           error: "User not found.",
         };
       }
-      // check password with args.password
       const passwordOk = await bcrypt.compare(password, user.password);
-      console.log(passwordOk);
       if (!passwordOk) {
         return {
           ok: false,
-          error: "Password incorrect.",
+          error: "Incorrect password.",
         };
       }
-      // issue a token and send it to the user
       const token = await jwt.sign({ id: user.id }, process.env.SECRET_KEY);
       return {
         ok: true,
