+++ users/login/login.typeDefs.js
@@ -1,0 +1,10 @@
+export default `#graphql
+  type LoginResult {
+    ok: Boolean!
+    token: String
+    error: String
+  }
+  type Mutation {
+    login(username: String!, password: String!): LoginResult!
+  }    
+`;
