+++ users/seeProfile/seeProfile.resolvers.js
@@ -1,0 +1,11 @@
+import client from "../../client.mjs";
+export default {
+  Query: {
+    seeProfile: (_, { username }) =>
+      client.user.findUnique({
+        where: {
+          username,
+        },
+      }),
+  },
+};
