+++ users/seeProfile/seeProfile.typeDefs.js
@@ -1,0 +1,5 @@
+export default `#graphql
+  type Query {
+    seeProfile(username: String!): User
+  }
+`;
