+++ users/users.typeDefs.js
@@ -8,22 +8,4 @@ export default `#graphql
     createdAt: String!
     updatedAt: String!
   }
-  type LoginResult {
-    ok: Boolean!
-    token: String
-    error: String
-  }
-  type Mutation {
-    createAccount(
-        firstName: String!
-        lastName: String
-        username: String!
-        email: String!
-        password: String!
-    ): User
-    login(username: String!, password: String!): LoginResult!
-  }
-  type Query{
-    seeProfile(username: String!): User
-  }
 `;
